Move the habit card to the widgets folder
